requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
langchain_community
youtube-transcript-api
pytube
python-dotenv
boto3
llama-parse
llama-index-core
llama-index-readers-file
pydantic
markitdown[all]
playwright
prometheus-client
pytest
pytest-asyncio
moto
Pillow
